# Conflicts:
#	mindocr/data/README.md
#	mindocr/data/det_dataset.py
#	mindocr/data/transforms/det_transforms.py
#	mindocr/data/transforms/general_transforms.py
#	mindocr/data/transforms/iaa_augment.py
#	mindocr/data/transforms/transforms_factory.py
#	tools/dataset_converters/convert.py
#	tools/dataset_converters/ctw1500.py
#	tools/dataset_converters/ic15.py
#	tools/dataset_converters/mlt2017.py
#	tools/dataset_converters/totaltext.py

+++ mindocr/data/transforms/general_transforms.py
@@ -4,12 +4,20 @@ from typing import List, Union
 import cv2
 import numpy as np
 from PIL import Image
+
 from mindspore.dataset.vision import RandomColorAdjust as MSRandomColorAdjust, ToPIL
 
 from ...data.constants import IMAGENET_DEFAULT_MEAN, IMAGENET_DEFAULT_STD
 
-__all__ = ['DecodeImage', 'NormalizeImage', 'ToCHWImage', 'PackLoaderInputs', 'RandomScale', 'RandomColorAdjust',
-           'RandomRotate', 'RandomHorizontalFlip']
+__all__ = [
+    "DecodeImage",
+    "NormalizeImage",
+    "ToCHWImage",
+    "PackLoaderInputs",
+    "RandomScale",
+    "RandomColorAdjust",
+    "RandomRotate",
+    "RandomHorizontalFlip"]
 
 
 class DecodeImage:
@@ -17,8 +25,10 @@ class DecodeImage:
     img_mode (str): The channel order of the output, 'BGR' and 'RGB'. Default to 'BGR'.
     channel_first (bool): if True, image shpae is CHW. If False, HWC. Default to False
     """
-    def __init__(self, img_mode='BGR', channel_first=False, to_float32=False, ignore_orientation=False, keep_ori=False,
-                 **kwargs):
+
+    def __init__(
+        self, img_mode="BGR", channel_first=False, to_float32=False, ignore_orientation=False, keep_ori=False, **kwargs
+    ):
         self.img_mode = img_mode
         self.to_float32 = to_float32
         self.channel_first = channel_first
@@ -26,28 +36,28 @@ class DecodeImage:
         self.keep_ori = keep_ori
 
     def __call__(self, data):
-        if 'img_path' in data:
-            with open(data['img_path'], 'rb') as f:
+        if "img_path" in data:
+            with open(data["img_path"], "rb") as f:
                 img = f.read()
-        elif 'img_lmdb' in data:
+        elif "img_lmdb" in data:
             img = data["img_lmdb"]
-        img = np.frombuffer(img, dtype='uint8')
+        img = np.frombuffer(img, dtype="uint8")
         img = cv2.imdecode(img, self.flag)
 
-        if self.img_mode == 'RGB':
+        if self.img_mode == "RGB":
             img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
 
         if self.channel_first:
             img = img.transpose((2, 0, 1))
 
         if self.to_float32:
-            img = img.astype('float32')
-        data['image'] = img
+            img = img.astype("float32")
+        data["image"] = img
         # data['ori_image'] = img.copy()
-        data['raw_img_shape'] = img.shape[:2]
+        data["raw_img_shape"] = img.shape[:2]
 
         if self.keep_ori:
-            data['image_ori'] = img.copy()
+            data["image_ori"] = img.copy()
 
         return data
 
@@ -58,8 +68,16 @@ class NormalizeImage:
     input image: by default, np.uint8, [0, 255], HWC format.
     return image: float32 numpy array
     """
-    def __init__(self, mean: Union[List[float], str] = 'imagenet', std: Union[List[float], str] = 'imagenet',
-                 is_hwc=True, bgr_to_rgb=False, rgb_to_bgr=False, **kwargs):
+
+    def __init__(
+        self,
+        mean: Union[List[float], str] = "imagenet",
+        std: Union[List[float], str] = "imagenet",
+        is_hwc=True,
+        bgr_to_rgb=False,
+        rgb_to_bgr=False,
+        **kwargs,
+    ):
         # By default, imagnet MEAN and STD is in RGB order. inverse if input image is in BGR mode
         self._channel_conversion = False
         if bgr_to_rgb or rgb_to_bgr:
@@ -67,12 +85,12 @@ class NormalizeImage:
 
         # TODO: detect hwc or chw automatically
         shape = (3, 1, 1) if not is_hwc else (1, 1, 3)
-        self.mean = np.array(self._get_value(mean, 'mean')).reshape(shape).astype('float32')
-        self.std = np.array(self._get_value(std, 'std')).reshape(shape).astype('float32')
+        self.mean = np.array(self._get_value(mean, "mean")).reshape(shape).astype("float32")
+        self.std = np.array(self._get_value(std, "std")).reshape(shape).astype("float32")
         self.is_hwc = is_hwc
 
     def __call__(self, data):
-        img = data['image']
+        img = data["image"]
         if isinstance(img, Image.Image):
             img = np.array(img)
         assert isinstance(img, np.ndarray), "invalid input 'img' in NormalizeImage"
@@ -83,18 +101,18 @@ class NormalizeImage:
             else:
                 img = img[[2, 1, 0], ...]
 
-        data['image'] = (img.astype('float32') - self.mean) / self.std
+        data["image"] = (img.astype("float32") - self.mean) / self.std
         return data
 
     @staticmethod
     def _get_value(val, name):
-        if isinstance(val, str) and val.lower() == 'imagenet':
-            assert name in ['mean', 'std']
-            return IMAGENET_DEFAULT_MEAN if name == 'mean' else IMAGENET_DEFAULT_STD
+        if isinstance(val, str) and val.lower() == "imagenet":
+            assert name in ["mean", "std"]
+            return IMAGENET_DEFAULT_MEAN if name == "mean" else IMAGENET_DEFAULT_STD
         elif isinstance(val, list):
             return val
         else:
-            raise ValueError(f'Wrong {name} value: {val}')
+            raise ValueError(f"Wrong {name} value: {val}")
 
 
 class ToCHWImage:
@@ -103,10 +121,10 @@ class ToCHWImage:
         pass
 
     def __call__(self, data):
-        img = data['image']
+        img = data["image"]
         if isinstance(img, Image.Image):
             img = np.array(img)
-        data['image'] = img.transpose((2, 0, 1))
+        data["image"] = img.transpose((2, 0, 1))
         return data
 
 
@@ -119,13 +137,14 @@ class PackLoaderInputs:
         input: data dict
         output: data tuple corresponding to the `output_columns`
     """
+
     def __init__(self, output_columns: List, **kwargs):
         self.output_columns = output_columns
 
     def __call__(self, data):
         out = []
         for k in self.output_columns:
-            assert k in data, f'key {k} does not exists in data, availabe keys are {data.keys()}'
+            assert k in data, f"key {k} does not exists in data, availabe keys are {data.keys()}"
             out.append(data[k])
 
         return tuple(out)
@@ -138,10 +157,11 @@ class RandomScale:
         scale_range: (min, max) scale range.
         p: probability of the augmentation being applied to an image.
     """
+
     def __init__(self, scale_range: Union[tuple, list], p: float = 0.5, **kwargs):
         self._range = scale_range
         self._p = p
-        assert kwargs.get('is_train', True), ValueError('RandomScale augmentation must be used for training only')
+        assert kwargs.get("is_train", True), ValueError("RandomScale augmentation must be used for training only")
 
     def __call__(self, data: dict):
         """
@@ -154,17 +174,17 @@ class RandomScale:
         """
         if random.random() < self._p:
             scale = np.random.uniform(*self._range)
-            data['image'] = cv2.resize(data['image'], dsize=None, fx=scale, fy=scale)
-            if 'polys' in data:
-                data['polys'] *= scale
+            data["image"] = cv2.resize(data["image"], dsize=None, fx=scale, fy=scale)
+            if "polys" in data:
+                data["polys"] *= scale
 
         return data
 
 
 class RandomColorAdjust:
     def __init__(self, brightness=32.0 / 255, saturation=0.5, **kwargs):
-        contrast = kwargs.get('contrast', (1, 1))
-        hue = kwargs.get('hue', (0, 0))
+        contrast = kwargs.get("contrast", (1, 1))
+        hue = kwargs.get("hue", (0, 0))
         self._jitter = MSRandomColorAdjust(brightness=brightness, saturation=saturation, contrast=contrast, hue=hue)
         self._pil = ToPIL()
 
@@ -174,7 +194,7 @@ class RandomColorAdjust:
         modified keys: image
         """
         # there's a bug in MindSpore that requires images to be converted to the PIL format first
-        data['image'] = np.array(self._jitter(self._pil(data['image'])))
+        data["image"] = np.array(self._jitter(self._pil(data["image"])))
         return data
